+++ src/main/gen/io/vlang/lang/psi/VlangCompileTimeMatchExpression.java
@@ -1,0 +1,45 @@
+// This is a generated file. Not intended for manual editing.
+package io.vlang.lang.psi;
+
+import java.util.List;
+import org.jetbrains.annotations.*;
+import com.intellij.psi.PsiElement;
+
+public interface VlangCompileTimeMatchExpression extends VlangExpression {
+
+  @Nullable
+  VlangCompileTimeMatchArms getCompileTimeMatchArms();
+
+  @Nullable
+  VlangExpression getExpression();
+
+  @Nullable
+  PsiElement getLbrace();
+
+  @NotNull
+  PsiElement getMatchCompileTime();
+
+  @Nullable
+  PsiElement getRbrace();
+
+  //WARNING: getArms(...) is skipped
+  //matching getArms(VlangCompileTimeMatchExpression, ...)
+  //methods are not found in VlangPsiImplUtil
+
+  //WARNING: getExpressionArms(...) is skipped
+  //matching getExpressionArms(VlangCompileTimeMatchExpression, ...)
+  //methods are not found in VlangPsiImplUtil
+
+  //WARNING: getTypeArms(...) is skipped
+  //matching getTypeArms(VlangCompileTimeMatchExpression, ...)
+  //methods are not found in VlangPsiImplUtil
+
+  //WARNING: getElseArm(...) is skipped
+  //matching getElseArm(VlangCompileTimeMatchExpression, ...)
+  //methods are not found in VlangPsiImplUtil
+
+  //WARNING: withElse(...) is skipped
+  //matching withElse(VlangCompileTimeMatchExpression, ...)
+  //methods are not found in VlangPsiImplUtil
+
+}