+++ src/main/gen/io/vlang/lang/psi/VlangSendStatement.java
@@ -8,13 +8,31 @@ import com.intellij.psi.PsiElement;
 public interface VlangSendStatement extends VlangStatement {
 
   @Nullable
+  VlangConstDeclaration getConstDeclaration();
+
+  @Nullable
+  VlangEnumDeclaration getEnumDeclaration();
+
+  @Nullable
   VlangExpression getExpression();
+
+  @Nullable
+  VlangFunctionDeclaration getFunctionDeclaration();
+
+  @Nullable
+  VlangGlobalVariableDeclaration getGlobalVariableDeclaration();
 
   @Nullable
   VlangImportList getImportList();
 
   @Nullable
+  VlangInterfaceDeclaration getInterfaceDeclaration();
+
+  @Nullable
   VlangLabelDefinition getLabelDefinition();
+
+  @Nullable
+  VlangMethodDeclaration getMethodDeclaration();
 
   @Nullable
   VlangModuleClause getModuleClause();
@@ -30,6 +48,15 @@ public interface VlangSendStatement extends VlangStatement {
 
   @Nullable
   VlangStatement getStatement();
+
+  @Nullable
+  VlangStaticMethodDeclaration getStaticMethodDeclaration();
+
+  @Nullable
+  VlangStructDeclaration getStructDeclaration();
+
+  @Nullable
+  VlangTypeAliasDeclaration getTypeAliasDeclaration();
 
   @NotNull
   PsiElement getSendChannel();