+++ src/main/gen/io/vlang/lang/psi/impl/VlangCompileTimeMatchExpressionImpl.java
@@ -1,0 +1,60 @@
+// This is a generated file. Not intended for manual editing.
+package io.vlang.lang.psi.impl;
+
+import java.util.List;
+import org.jetbrains.annotations.*;
+import com.intellij.lang.ASTNode;
+import com.intellij.psi.PsiElement;
+import com.intellij.psi.PsiElementVisitor;
+import io.vlang.lang.psi.VlangPsiTreeUtil;
+import static io.vlang.lang.VlangTypes.*;
+import io.vlang.lang.psi.*;
+
+public class VlangCompileTimeMatchExpressionImpl extends VlangExpressionImpl implements VlangCompileTimeMatchExpression {
+
+  public VlangCompileTimeMatchExpressionImpl(@NotNull ASTNode node) {
+    super(node);
+  }
+
+  @Override
+  public void accept(@NotNull VlangVisitor visitor) {
+    visitor.visitCompileTimeMatchExpression(this);
+  }
+
+  @Override
+  public void accept(@NotNull PsiElementVisitor visitor) {
+    if (visitor instanceof VlangVisitor) accept((VlangVisitor)visitor);
+    else super.accept(visitor);
+  }
+
+  @Override
+  @Nullable
+  public VlangCompileTimeMatchArms getCompileTimeMatchArms() {
+    return VlangPsiTreeUtil.getChildOfType(this, VlangCompileTimeMatchArms.class);
+  }
+
+  @Override
+  @Nullable
+  public VlangExpression getExpression() {
+    return VlangPsiTreeUtil.getChildOfType(this, VlangExpression.class);
+  }
+
+  @Override
+  @Nullable
+  public PsiElement getLbrace() {
+    return findChildByType(LBRACE);
+  }
+
+  @Override
+  @NotNull
+  public PsiElement getMatchCompileTime() {
+    return notNullChild(findChildByType(MATCH_COMPILE_TIME));
+  }
+
+  @Override
+  @Nullable
+  public PsiElement getRbrace() {
+    return findChildByType(RBRACE);
+  }
+
+}